Fixed text cropping errors by double casting

+++ BlockBuddies Main/BlockBuddies/GraphicsManager.cpp
@@ -84,8 +84,8 @@ sf::Vector2f GraphicsManager::getCenter(const sf::Text& text, Bounds bounds)
 	else if (bounds == Bounds::GLOBAL)
 		textRect = text.getGlobalBounds();
 
-	return sf::Vector2f(textRect.left + textRect.width/2.0f,
-		                textRect.top + textRect.height/2.0f);
+	return sf::Vector2f((float)(int)(textRect.left + textRect.width/2.0f),
+                        (float)(int)(textRect.top + textRect.height / 2.0f));
 }
 
 // Basically does the same thing as getCenter except only applies any
@@ -99,8 +99,8 @@ sf::Vector2f GraphicsManager::getLeftCenter(const sf::Text& text, Bounds bounds)
 	else if (bounds == Bounds::GLOBAL)
 		textRect = text.getGlobalBounds();
 
-	return sf::Vector2f(textRect.left,
-		                textRect.top + textRect.height / 2.0f);
+	return sf::Vector2f((float)(int)(textRect.left),
+                        (float)(int)(textRect.top + textRect.height / 2.0f));
 }
 
 // Finds the right center point of a Text object
@@ -113,8 +113,8 @@ sf::Vector2f GraphicsManager::getRightCenter(const sf::Text& text, Bounds bounds
 	else if (bounds == Bounds::GLOBAL)
 		textRect = text.getGlobalBounds();
 
-	return sf::Vector2f(textRect.left + textRect.width,
-		                textRect.top + textRect.height / 2.0f);
+	return sf::Vector2f((float)(int)(textRect.left + textRect.width),
+                        (float)(int)(textRect.top + textRect.height / 2.0f));
 }
 
 // Overloaded getCenter() function for sf::RectangleShape
@@ -126,9 +126,9 @@ sf::Vector2f GraphicsManager::getCenter(const sf::RectangleShape& rectangle, Bou
 		rect = rectangle.getLocalBounds();
 	else if (bounds == Bounds::GLOBAL)
 		rect = rectangle.getGlobalBounds();
-
-	return sf::Vector2f(rect.left + rect.width/2.0f,
-                        rect.top + rect.height/2.0f);		
+    
+	return sf::Vector2f((float)(int)(rect.left + rect.width/2.0f),
+                        (float)(int)(rect.top + rect.height/2.0f));		
 }
 
 // Returns the left-center point of a rectangle
@@ -141,8 +141,8 @@ sf::Vector2f GraphicsManager::getLeftCenter(const sf::RectangleShape& rectangle,
 	else if (bounds == Bounds::GLOBAL)
 		rect = rectangle.getGlobalBounds();
 
-	return sf::Vector2f(rect.left,
-                        rect.top + rect.height/2.0f);		
+	return sf::Vector2f((float)(int)(rect.left),
+                        (float)(int)(rect.top + rect.height/2.0f));		
 }
 
 // Finds the right center point of a rectangle
@@ -155,6 +155,6 @@ sf::Vector2f GraphicsManager::getRightCenter(const sf::RectangleShape& rectangle
 	else if (bounds == Bounds::GLOBAL)
 		rect = rectangle.getGlobalBounds();
 
-	return sf::Vector2f(rect.left + rect.width,
-                        rect.top + rect.height/2.0f);		
+	return sf::Vector2f((float)(int)(rect.left + rect.width),
+                        (float)(int)(rect.top + rect.height/2.0f));
 }
